fix(admin-products): atomic reorder + manual slug override + review fixes
- FIX 1: wrap reorderImages sort_order UPDATE loop in DB::transaction (AM-2)
- FIX 2: add optional slug field to UpdateProductRequest with unique rule;
  controller uses manual slug directly when provided (AM-3); +2 tests (RED→GREEN)
- FIX 3: sort_order column changed to unsignedTinyInteger to mirror service_images
- FIX 4: remove redundant explicit index('category_id') on products table
  (foreignId constrained already creates one)
- FIX 5: ProductDetailResource images_count uses null-coalesce pattern
  ($this->images_count ?? $this->images->count()) to match ProductCardResource
- FIX 6: add student 403 tests for update/reorder/destroyImage; add 9+1=10
  boundary test (total-at-cap accepted); +4 tests
- FIX 7: reorder route changed PATCH → POST (spec AM-2); test updated to
  postJson; add counter-starts-at-2 comment in uniqueSlug per AM-3

+++ backend/database/migrations/2026_06_19_900001_create_product_images_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
                   ->constrained()
                   ->cascadeOnDelete();
             $table->string('path');
-            $table->unsignedInteger('sort_order')->default(0);
+            $table->unsignedTinyInteger('sort_order')->default(0);
             $table->timestamps();
         });
     }
